requirements.txt
streamlit==1.35.0
tensorflow==2.15.0
Pillow==10.0.0
numpy>=1.22.0
